+++ firebase-vscode/src/data-connect/emulator-stream.ts
@@ -3,6 +3,7 @@ import fetch from "node-fetch";
 import { Observable, of } from "rxjs";
 import { backOff } from "exponential-backoff";
 import { ResolvedDataConnectConfigs } from "./config";
+import { Signal } from "@preact/signals-core";
 
 enum Kind {
   KIND_UNSPECIFIED = "KIND_UNSPECIFIED",
@@ -28,19 +29,20 @@ export const emulatorOutputChannel =
   vscode.window.createOutputChannel("Firebase Emulators");
 
 /**
- *
+ * TODO: convert to class
  * @param fdcEndpoint FDC Emulator endpoint
  */
 export async function runEmulatorIssuesStream(
   configs: ResolvedDataConnectConfigs,
   fdcEndpoint: string,
+  isPostgresEnabled: Signal<boolean>,
 ) {
   const obsErrors = await getEmulatorIssuesStream(configs, fdcEndpoint);
   const obsConverter = {
     next(nextResponse: EmulatorIssueResponse) {
       if (nextResponse.result?.issues?.length) {
         for (const issue of nextResponse.result.issues) {
-          displayIssue(issue);
+          displayAndHandleIssue(issue, isPostgresEnabled);
         }
       }
     },
@@ -57,7 +59,7 @@ export async function runEmulatorIssuesStream(
 /**
  * Based on the severity of the issue, either log, display notification, or display interactive popup to the user
  */
-export function displayIssue(issue: EmulatorIssue) {
+export function displayAndHandleIssue(issue: EmulatorIssue, isPostgresEnabled: Signal<boolean>) {
   const issueMessage = `Data Connect Emulator: ${issue.kind.toString()} - ${issue.message}`;
   if (issue.severity === Severity.ALERT) {
     vscode.window.showErrorMessage(issueMessage);
@@ -65,6 +67,11 @@ export function displayIssue(issue: EmulatorIssue) {
     vscode.window.showWarningMessage(issueMessage);
   }
   emulatorOutputChannel.appendLine(issueMessage);
+
+  // special handlings
+  if (issue.kind === Kind.SQL_CONNECTION) {
+    isPostgresEnabled.value = false;
+  }
 }
 
 /**