+++ src/emulator/controller.ts
@@ -821,7 +821,6 @@ export async function startAll(
   }
 
   if (listenForEmulator.dataconnect) {
-    const dataConnectAddr = legacyGetFirstAddr(Emulators.DATACONNECT);
     const config = readFirebaseJson(options.config);
     if (!config.length) {
       throw new FirebaseError("No Data Connect service found in firebase.json");
@@ -837,8 +836,7 @@ export async function startAll(
     }
     const localConnectionString = getLocalConectionString(options.rc);
     const dataConnectEmulator = new DataConnectEmulator({
-      host: dataConnectAddr.host,
-      port: dataConnectAddr.port,
+      listen: listenForEmulator.dataconnect,
       projectId,
       auto_download: true,
       configDir,
@@ -846,6 +844,9 @@ export async function startAll(
       localConnectionString,
     });
     await startEmulator(dataConnectEmulator);
+    if (!utils.isVSCodeExtension()) {
+      await dataConnectEmulator.connectToPostgres();
+    }
   }
 
   if (listenForEmulator.storage) {